Intermediate refactoring work
