+++ src/dipProof/extensions/index.ts
@@ -5,10 +5,4 @@
  * found in the LICENSE file in the root directory of this source tree.
  */
 
-/**
- * @module @kiltprotocol/dip-sdk
- */
-
-export * as stateProof from './stateProof/index.js'
-export * as dipProof from './dipProof/index.js'
-export * from "./sibling.js"
+export * from './timeBoundDidSignature.js'
